requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
Pillow==10.2.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.34
django-cleanup==8.0.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-environ==0.11.2
django-allauth==0.60.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
drf-yasg==1.21.7 